feat: use extension, document
